% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{convergence}
\alias{convergence}
\title{MCMC convergence diagnostics}
\usage{
convergence(lg, plots = TRUE)
}
\arguments{
\item{lg}{Output from a call to \code{lgcp}}

\item{plots}{Logical indicating whether to plot MCMC traceplots}
}
\value{
Traceplot of the MCMC chains of parameters from the linear predictor and
covariance function is plotted, and R-hat and ESS statistics are printed.
}
\description{
Diagnostics for convergence of the MCMC chains from a call to \code{lgcp}.
}
\details{
Produces a traceplot of the model parameters and prints R-hat and ESS statistics.
}
