% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Simulated point process data set}
\format{
A data frame with 183 rows and 3 variables:
\describe{
\item{x}{the x coordinate of the point}
\item{y}{the y coordinate of the point}
\item{t}{the time period of the point}
}
}
\usage{
dat
}
\description{
A simulated dataset with seven time periods over a unit square with population
density given by \code{square_pop}. Primarily used for testing.
}
\keyword{datasets}
