% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{.lgcpST}
\alias{.lgcpST}
\title{Alternative to lgcpPredictSpatioTemporalPlusPars}
\usage{
.lgcpST(
  formula,
  xyt,
  T,
  laglength,
  ZmatList = NULL,
  model.priors,
  model.inits = lgcpInits(),
  spatial.covmodel,
  cellwidth = NULL,
  poisson.offset = NULL,
  mcmc.control,
  output.control = setoutput(),
  gradtrunc = Inf,
  ext = 2,
  inclusion = "touching"
)
}
\description{
Alternative to lgcpPredictSpatioTemporalPlusPars
}
