% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{.plot_lgcp_dat}
\alias{.plot_lgcp_dat}
\title{Extract data from an lgcpReal object}
\usage{
.plot_lgcp_dat(
  samps,
  grid.data,
  lg,
  nchains,
  idx.mapping = NULL,
  cellwidth = 0.005,
  covariates,
  plotlag = 0,
  data.out = FALSE
)
}
\description{
Extract data from an lgcpReal object
}
