% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{generateReport}
\alias{generateReport}
\title{Report generating function}
\usage{
generateReport(
  lg,
  plot.opts = list(),
  dirname = getwd(),
  change.lag = NULL,
  par.summ = FALSE,
  hotspot.opts = NULL,
  breakdown.plots = TRUE,
  aggregate = NULL,
  repdate = format(Sys.time(), "\%d \%B, \%Y"),
  add3D = FALSE,
  pdf = FALSE
)
}
\description{
To be completed
}
