% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get_day}
\alias{get_day}
\title{Get day of the week}
\usage{
get_day(df)
}
\arguments{
\item{df}{A data frame containing columns with the address (either "Address" or "address") and date (either "Date" or "date")
of the cases. Typically the same data frame used in a call to \code{geocode_st}}
}
\value{
A data frame with value for time period and day of the week.
}
\description{
Returns a data frame with day of the week covariates
}
\examples{
\dontrun{
tmp <- data.frame(address=c("Buckingham palace","Big ben, Westminster","Marble arch, London"),
                  date = c("01/01/2020","02/01/2020","03/01/2020"))
get_day(tmp)
}
}
