% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.lgcpRealPlot}
\alias{plot.lgcpRealPlot}
\title{(Re-)Plot lgcpRealPlot objects}
\usage{
\method{plot}{lgcpRealPlot}(obj)
}
\arguments{
\item{obj}{Output from a previous call to plotting functions}
}
\description{
Plot lgcpRealPlot output from plot, plot_hotspot, aggregator.
}
\examples{
plot(p1)
}
