% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.lgcpRealPlot}
\alias{print.lgcpRealPlot}
\title{Print function for lgcpRealPlot}
\usage{
\method{print}{lgcpRealPlot}(x, ...)
}
\arguments{
\item{x}{An lgcpRealPlot object}

\item{...}{...}
}
\value{
Plots the object which is a list of ggplot2 plots.
}
\description{
Print function for lgcpRealPlot
}
