% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.lgcpRealSumm}
\alias{print.lgcpRealSumm}
\title{Print function for lgcpRealSumm}
\usage{
\method{print}{lgcpRealSumm}(x, ..., digits = 3)
}
\arguments{
\item{x}{An lgcpRealSumm object}

\item{...}{...}

\item{digits}{Number of digits to round the summary to}
}
\value{
A summary of the output of an lgcp model.
}
\description{
Print function for lgcpRealSumm
}
