% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{square}
\alias{square}
\title{A unit square spatial polygon}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 1 columns.
}
\usage{
square
}
\description{
A SpatialPolygonsDataFrame describing a unit square
}
\keyword{datasets}
