% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{square_pop}
\alias{square_pop}
\title{Simulated population density across unit square}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 121 rows and 2 columns.
}
\usage{
square_pop
}
\description{
A SpatialPolygonsDataFrame covering a unit square with 11x11 square polygons, each with a
different simulated population density. Used to simulate the data in \code{dat}
}
\keyword{datasets}
