% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{summary.lgcpReal}
\alias{summary.lgcpReal}
\title{Summarise lgcp output}
\usage{
\method{summary}{lgcpReal}(object, linear = TRUE, plot = TRUE, prior = FALSE)
}
\description{
Summarise lgcp output
}
