% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{aggregator_data}
\alias{aggregator_data}
\title{Create spatial polygons data frame}
\usage{
aggregator_data(obj, aggpoly, osm = FALSE)
}
\arguments{
\item{obj}{A lgcpReal object from the lgcp function}

\item{aggpoly}{A \code{spatialPolygons} or \code{spatialPolygonsDataFrame} describing the
geography to aggregate to.}
}
\value{
A \code{spatialPolygonsDataFrame}
}
\description{
Aggregate to larger geography and output data
}
\details{
This function aggregates the output of plot or plot_hotspot to a
larger geography and outputs a spatial polygons data frame that
can be used to create other plots.
}
