% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{get_map2}
\alias{get_map2}
\title{Alternative \code{get_map} function}
\usage{
get_map2(
  location = c(lon = -95.3632715, lat = 29.7632836),
  zoom = "auto",
  scale = "auto",
  maptype = c("terrain", "terrain-background", "satellite", "roadmap", "hybrid",
    "toner", "watercolor", "terrain-labels", "terrain-lines", "toner-2010", "toner-2011",
    "toner-background", "toner-hybrid", "toner-labels", "toner-lines", "toner-lite"),
  source = c("google", "osm", "stamen"),
  force = ifelse(source == "google", TRUE, FALSE),
  messaging = FALSE,
  urlonly = FALSE,
  filename = NULL,
  crop = TRUE,
  color = c("color", "bw"),
  language = "en-EN",
  ...
)
}
\description{
An alternative to \code{ggmap}'s \code{get_map} function
}
\details{
An error in the CRAN available version of get_map means it will only plot Google Maps objects rather than Stamen or OSM maps. When ggmap is
updated this function will be removed. See \code{help(get_map)} for details.
}
