% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_hotspot}
\alias{plot_hotspot}
\title{Hotspot mapping and visualisation}
\usage{
plot_hotspot(
  lg1,
  covariates,
  threshold.var = NULL,
  threshold.value = NULL,
  labels,
  threshold.prob = 0.8,
  relative = TRUE,
  osm = FALSE,
  per.days = 10000,
  msq = 10000
)
}
\description{
A function for mapping hotspots according to user defined criteria.
}
\details{
A ``hotspot'' is defined as an area that exceeds a user-defined criterion with
probability of at least p. The criterion can be a function of one or two variables
derived from the model; where two variables are used then there are four possible
hotspot classifications, where only one is used then there are two classifications
(above or below the threshold).
}
