% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.lgcpReal}
\alias{print.lgcpReal}
\title{Print function for lgcpReal}
\usage{
\method{print}{lgcpReal}(x, ...)
}
\arguments{
\item{x}{An lgcpRealPlot object}

\item{...}{...}
}
\value{
A string summarising the model
}
\description{
Print function for lgcpReal
}
