% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{summary.lgcpReal}
\alias{summary.lgcpReal}
\title{Summarise lgcp output}
\usage{
summary(lg,...)
}
\arguments{
\item{object}{An \code{lgcpReal} object from a call to \code{lgcp}}

\item{linear}{A logical value indicating whether results should be reported on linear or exponential scales}

\item{plot}{A logical value indicating whether to produce plots of the prior and posterior distributions of model parameters}

\item{prior}{A logical value indicating whether to report prior distributions in the summary output}
}
\value{
A table with posterior mean, SD, and quantiles of posterior and prior distributions.
}
\description{
Summarise output from a call to \code{lgcp}
}
